Build_SQL_AI_AGENTS_WITH_LLM_Models/requirements.txt
openai
langchain
sqlalchemy
rich
python-dotenv
